[Android] Loading progress

+++ common/src/main/kotlin/io/github/dector/knight/common/TimeUtils.kt
@@ -9,6 +9,6 @@ fun Date.add(time: Long, unit: TimeUnit) = Date(this.time + unit.toMillis(time))
 
 // ----- Time processing -----
 
-fun Int.secondsAsMillis(): Int = this * 1000
+fun Int.secondsAsMillis(): Long = this * 1000L
 
-fun Int.minutesAsMillis(): Int = this * 60.secondsAsMillis()
+fun Int.minutesAsMillis(): Long = this * 60.secondsAsMillis()
